fixed some errors with the authprovider
