+++ app/page.tsx
@@ -17,11 +17,14 @@ import {
   removeClientCookie,
 } from '@/app/utils/stores/cookieStore';
 import { Skeleton } from '@/components/ui/skeleton';
+import { useAuth } from "@/app/user/auth/context/AuthContext";
 
 /// Main card displayed with login buttons etc
 function HeroCard() {
   const [isLoggedIn, setIsLoggedIn] = useState(false);
   const [isLoading, setIsLoading] = useState(true);
+
+  const {logout} = useAuth();
 
   useEffect(() => {
     checkIfLoggedIn();
@@ -35,6 +38,7 @@ function HeroCard() {
 
   const handleLogout = () => {
     setIsLoggedIn(false);
+    logout();
     removeClientCookie('tokenDrinking');
   };
 
@@ -107,7 +111,7 @@ function HeroCard() {
           <a
             href="https://www.instagram.com/tihlde/"
             target="_blank"
-            className="flex justify-center"
+            className="flex justify-center" rel="noreferrer"
           >
             <Button variant={'outline'} size={'icon'}>
               <Instagram />
@@ -117,7 +121,7 @@ function HeroCard() {
           <a
             href="https://twitter.com/tihlde"
             target="_blank"
-            className="flex justify-center"
+            className="flex justify-center" rel="noreferrer"
           >
             <Button variant={'outline'} size={'icon'}>
               <Twitter />
