+++ app/user/auth/context/AuthContext.js
@@ -1,27 +1,34 @@
 "use client";
 
-import React, { createContext, useContext, useState, useEffect } from 'react';
+import React, { createContext, useContext, useState, useLayoutEffect } from 'react';
 import { getClientCookie } from '@/app/utils/stores/cookieStore';
-const AuthContext = createContext(null);
 
+const AuthContext = createContext({
+  isLoggedIn: false,
+  login: () => {},
+  logout: () => {}
+});
 
 export const AuthProvider = ({ children }) => {
   const [isLoggedIn, setIsLoggedIn] = useState(false);
 
-
-  useEffect(() => {
-    checkLoginStatus();
-  }, []);
-
-  const checkLoginStatus = () => {
+  useLayoutEffect(() => {
     const token = getClientCookie('tokenDrinking');
     if (token) {
       setIsLoggedIn(true);
     }
+  }, []);
+
+  const login = () => {
+    setIsLoggedIn(true);
+  };
+
+  const logout = () => {
+    setIsLoggedIn(false);
   };
 
   return (
-    <AuthContext.Provider value={{ isLoggedIn, setIsLoggedIn }}>
+    <AuthContext.Provider value={{ isLoggedIn, login, logout }}>
       {children}
     </AuthContext.Provider>
   );
